refactor: changes to RANSAC, PnP and BA
RANSAC:
1. Change SampleConcensusEstimator -> SampleConcensus
2. RefineModel() add error threshold paramter
3. SampleSize() return int

PnP:
1. Move PnP applications into PnP folder

BA:
1. Move Options and Summary into BundleAdjustment class
2. Move all the BA helpers under BundleAdjustment class
3. Add BA to estimate Homography
4. BA use impl pattern
5. BA add camera intrinsics covariance calculation

Others:
1. Put feature related class into one file
2. Changes because of all above
3. Formatting, renaming, and etc

+++ modules/math/tests/test_utils.h
@@ -32,7 +32,7 @@ public:
     LineEstimator() {}
     ~LineEstimator() {}
 
-    double SampleSize() const override { return 2; }
+    int SampleSize() const override { return 2; }
 
     bool EstimateModel(const std::vector<Point> &data,
                        std::vector<Line> *models) const override
